Refactor to use custon context hooks.
