+++ src/contexts/CurrentUserContext.js
@@ -1,3 +1,12 @@
+import { createContext, useContext, useEffect, useState } from "react";
+import axios from "axios";
+
+export const CurrentUserContext = createContext();
+export const SetCurrentUserContext = createContext();
+
+export const useCurrentUser = () => useContext(CurrentUserContext);
+export const useSetCurrentUser = () => useContext(SetCurrentUserContext);
+
 export const CurrentUserProvider = ({ children }) => {
     const [currentUser, setCurrentUser] = useState(null);
 
@@ -13,4 +22,12 @@ export const CurrentUserProvider = ({ children }) => {
     useEffect(() => {
         handleMount();
     }, []);
+
+    return (
+        <CurrentUserContext.Provider value={currentUser}>
+          <SetCurrentUserContext.Provider value={setCurrentUser}>
+            {children}
+          </SetCurrentUserContext.Provider>
+        </CurrentUserContext.Provider>
+      );
 };
