+++ src/index.js
@@ -4,11 +4,14 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { BrowserRouter as Router } from "react-router-dom";
+import { CurrentUserProvider } from './contexts/CurrentUserContext';
 
 ReactDOM.render(
   <React.StrictMode>
-        <Router>
-      <App />
+    <Router>
+      <CurrentUserProvider>
+        <App />
+      </CurrentUserProvider>
     </Router>
   </React.StrictMode>,
   document.getElementById('root')
